feat: 指定模块统一打包插件 #8672
