+++ src/backend/ci/core/repository/biz-repository/src/main/kotlin/com/tencent/devops/repository/resources/ServiceRepositoryGitCheckResourceImpl.kt
@@ -5,12 +5,12 @@ import com.tencent.devops.common.api.pojo.Result
 import com.tencent.devops.common.web.RestResource
 import com.tencent.devops.repository.api.ServiceRepositoryGitCheckResource
 import com.tencent.devops.repository.pojo.RepositoryGitCheck
-import com.tencent.devops.repository.service.RepoGitCheckService
+import com.tencent.devops.repository.service.GitCheckService
 import org.springframework.beans.factory.annotation.Autowired
 
 @RestResource
 class ServiceRepositoryGitCheckResourceImpl @Autowired constructor(
-    private val repoGitCheckService: RepoGitCheckService
+    private val gitCheckService: GitCheckService
 ) : ServiceRepositoryGitCheckResource {
     override fun getGitCheck(
         pipelineId: String,
@@ -18,14 +18,14 @@ class ServiceRepositoryGitCheckResourceImpl @Autowired constructor(
         context: String,
         repositoryConfig: RepositoryConfig
     ): Result<RepositoryGitCheck?> {
-        return Result(repoGitCheckService.getGitCheck(pipelineId, repositoryConfig, commitId, context))
+        return Result(gitCheckService.getGitCheck(pipelineId, repositoryConfig, commitId, context))
     }
 
     override fun createGitCheck(gitCheck: RepositoryGitCheck) {
-        repoGitCheckService.creatGitCheck(gitCheck)
+        gitCheckService.creatGitCheck(gitCheck)
     }
 
     override fun updateGitCheck(gitCheckId: Long, buildNumber: Int) {
-        repoGitCheckService.updateGitCheck(gitCheckId = gitCheckId, buildNumber = buildNumber)
+        gitCheckService.updateGitCheck(gitCheckId = gitCheckId, buildNumber = buildNumber)
     }
 }
