+++ src/backend/ci/core/repository/biz-repository/src/main/kotlin/com/tencent/devops/repository/service/GitCheckService.kt
@@ -9,7 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
 
 @Service
-class RepoGitCheckService @Autowired constructor(
+class GitCheckService @Autowired constructor(
     private val dslContext: DSLContext,
     private val gitCheckDao: GitCheckDao
 ) {
