+++ src/backend/ci/core/stream/biz-stream/src/main/kotlin/com/tencent/devops/stream/resources/external/ExternalStreamResourceImpl.kt
@@ -25,41 +25,27 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.stream.api.external
+package com.tencent.devops.stream.resources.external
 
-import io.swagger.annotations.Api
-import io.swagger.annotations.ApiOperation
-import io.swagger.annotations.ApiParam
-import javax.ws.rs.Consumes
-import javax.ws.rs.GET
-import javax.ws.rs.Path
-import javax.ws.rs.PathParam
-import javax.ws.rs.Produces
-import javax.ws.rs.QueryParam
-import javax.ws.rs.core.MediaType
+import com.tencent.devops.common.web.RestResource
+import com.tencent.devops.stream.api.external.ExternalStreamResource
+import com.tencent.devops.stream.service.StreamPipelineBadgeService
 
-@Api(tags = ["EXTERNAL_STREAM"], description = "外部-STREAM资源获取")
-@Path("/external/stream")
-@Produces(MediaType.APPLICATION_JSON)
-@Consumes(MediaType.APPLICATION_JSON)
-interface StreamExternalStreamResource {
-
-    @ApiOperation("获取流水线徽章")
-    @Produces("image/svg+xml")
-    @GET
-    @Path("/projects/{gitProjectId}/pipelines/badge")
-    fun getPipelineBadge(
-        @ApiParam("Git仓库ID", required = true)
-        @PathParam("gitProjectId")
+@RestResource
+class ExternalStreamResourceImpl(
+    private val streamPipelineBadgeService: StreamPipelineBadgeService
+) : ExternalStreamResource {
+    override fun getPipelineBadge(
         gitProjectId: Long,
-        @ApiParam("流水线文件名称", required = true)
-        @QueryParam("file_path")
         filePath: String,
-        @ApiParam("分支名称", required = false)
-        @QueryParam("branch")
         branch: String?,
-        @ApiParam("触发方式", required = false)
-        @QueryParam("object_kind")
         objectKind: String?
-    ): String
+    ): String {
+        return streamPipelineBadgeService.get(
+            gitProjectId = gitProjectId,
+            filePath = filePath,
+            branch = branch,
+            objectKind = objectKind
+        )
+    }
 }
